Move element definition normalize to separate file.
